requirements.txt
kafka-python
pandas
awswrangler
s3fs
yfinance
streamlit
altair
plotly_express